Move in StarLasuLWLanguage the logic to of its primitive types serializers and deserializers
