+++ lionweb/build.gradle
@@ -21,7 +21,7 @@ dependencies {
     testImplementation "org.jetbrains.kotlin:kotlin-test-junit:$kotlin_version"
     testImplementation "com.google.code.gson:gson:$gson_version"
     api(libs.lionwebjava)
-    api("io.lionweb.lionweb-kotlin:lionweb-kotlin-2024.1-core:0.2.2") {
+    api(libs.lionwebkotlincore) {
         exclude group: "org.jetbrains.kotlin", module: "kotlin-test-junit5"
     }
     api(project(":core"))
