+++ lionweb/src/test/kotlin/com/strumenta/kolasu/lionweb/LionWebModelConverterTest.kt
@@ -717,6 +717,7 @@ class LionWebModelConverterTest {
 
     @Test
     fun canDeserializePosition() {
+        StarLasuLWLanguage
         val kl = KolasuLanguage("my.language").apply {
             addClass(NodeWithEnum::class)
         }
